Refactored the Manifest Processor class and test
Had to refactor to align with the revised QTI Reader class. All that was changed was that the manifest processor class no longer handles the unzipping (cause the QTI Reader will handle it). But for the test to pass,  it has to unzip the QTI file before verifying the existence of the manifest file.